requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
asyncpg~=0.30.0